pyproject.toml
[project]
name = "portfolio-service"
version = "0.1.0"
description = "A Model Context Protocol server implementing popular broker api's"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
 "aiohttp>=3.11.8",
 "ib-insync>=0.9.86",
 "mcp>=1.0.0",
 "nest-asyncio>=1.6.0",
 "pandas>=2.2.3",
 "pydantic>=2.10.2",
 "python-dotenv>=1.0.1",
 "requests>=2.32.3",
 "retry>=0.9.2",
 "tradingview-screener>=2.5.0",
]
[[project.authors]]
name = "Itay Verkh"
email = "itay1542@gmail.com"

[build-system]
requires = [ "hatchling",]
build-backend = "hatchling.build"

[project.scripts]
portfolio-service = "portfolio_service:main"
